refactor:  io 제외한 모든 패키지 model 패키지로 이동
