+++ src/main/java/com/devcourse/engine/model/computer/SimpleComputer.java
@@ -1,4 +1,4 @@
-package com.devcourse.engine.computer;
+package com.devcourse.engine.model.computer;
 
 import com.devcourse.engine.model.Operator;
