+++ src/main/java/com/devcourse/engine/model/exception/InvalidInputException.java
@@ -1,4 +1,4 @@
-package com.devcourse.engine.exception;
+package com.devcourse.engine.model.exception;
 
 public class InvalidInputException extends RuntimeException {
