+++ src/test/java/engine/computer/ComputerTest.java
@@ -1,10 +1,10 @@
 package engine.computer;
 
-import com.devcourse.engine.computer.Computer;
-import com.devcourse.engine.computer.SimpleComputer;
-import com.devcourse.engine.converter.PostfixConverter;
-import com.devcourse.engine.exception.InvalidInputException;
-import com.devcourse.engine.validator.SimpleValidator;
+import com.devcourse.engine.model.computer.Computer;
+import com.devcourse.engine.model.computer.SimpleComputer;
+import com.devcourse.engine.model.converter.PostfixConverter;
+import com.devcourse.engine.model.exception.InvalidInputException;
+import com.devcourse.engine.model.validator.SimpleValidator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
