+++ src/test/java/engine/computer/ValidatorTest.java
@@ -1,8 +1,8 @@
 package engine.computer;
 
-import com.devcourse.engine.exception.InvalidInputException;
-import com.devcourse.engine.validator.SimpleValidator;
-import com.devcourse.engine.validator.Validator;
+import com.devcourse.engine.model.exception.InvalidInputException;
+import com.devcourse.engine.model.validator.SimpleValidator;
+import com.devcourse.engine.model.validator.Validator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
